Move models/tournament into models root and models/tournament/fixture into models/fixtures (#75)
