+++ public/components/game/details/details.js
@@ -33,7 +33,7 @@ require("can/route/");
 require("can/view/href/");
 
 var platform = require( "steal-platform" );
-var Tournament = require("bitballs/models/tournament/");
+var Tournament = require("bitballs/models/tournament");
 var Game = require("bitballs/models/game");
 var Team = require("bitballs/models/team");
 var Player = require("bitballs/models/player");
