+++ public/components/tournament/details/details.html
@@ -1,7 +1,7 @@
 <div id="main"></div>
 
 <script src="../../../node_modules/steal/steal.js" main="@empty">
-    import "bitballs/models/tournament/fixture";
+    import "bitballs/models/fixtures/tournament";
     import "bitballs/components/tournament/details/";
     import $ from 'jquery';
     import stache from "can/view/stache/";
