+++ public/components/tournament/details/viewModel.js
@@ -2,7 +2,7 @@ var Map = require("can/map/");
 var Team = require("bitballs/models/team");
 var Game = require("bitballs/models/game");
 var Player = require("bitballs/models/player");
-var Tournament = require("bitballs/models/tournament/");
+var Tournament = require("bitballs/models/tournament");
 
 require("can/map/define/");
